Updating `loop_translation` file to `expNN_BLAS_level_2_to_level_3`
